Add support for retrieving `Contribute` from readme
Previously, if there was no `CONTRIBUTE.md` file, the fixture in this
repo was used.  Now, in this case, and if there’s a `Contribute`
section on the README of the repository, that section (including the
heading) is used instead as a template for `CONTRIBUTE.md`.

Related to GH-12.
Supersedes GH-22.